Added ReadToEndOfFile to parser2
Added ReadToEndOfFile to parser2.c, which executes when the parser cannot resynchronise. When called, this will add an error message and remainder of source file to the output file. Called this function agter ParseProgram, rather than in Accept, where it was before, causing a segmentation error.

Also tidied up code a bit and improved the ReadToEndOfFile in parser1